+++ source/renderer/zettlr-rendereripc.js
@@ -112,9 +112,19 @@ class ZettlrRendererIPC {
     // Sends an array of IDs to main. If they are found in the JSON, cool! Otherwise
     // this will return false.
     global.citeproc = {
-      getCitation: (citation) => { return ipc.sendSync('getCitation', citation) },
-      updateItems: (keyList) => { return ipc.sendSync('updateItems', keyList) },
-      makeBibliography: () => { this.send('citeproc-make-bibliography') }
+      getCitation: (citation) => {
+        return ipc.sendSync('cite', {
+          'type': 'get-citation',
+          'content': citation
+        })
+      },
+      updateItems: (keyList) => {
+        return ipc.sendSync('cite', {
+          'type': 'update-items',
+          'content': keyList
+        })
+      },
+      makeBibliography: () => { ipc.send('cite', { 'type': 'make-bibliography' }) }
     }
 
     global.ipc = {