+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 
 - Massive rewrite of the underlying mechanism of loading the directory trees into the app. It's now asynchronous and the app starts up way faster than before.
 - Replaced the citation engine with `Citr` for more accurate results in previewing citations.
+- The `Citeproc` engine is now a service provider.
 
 # 1.3.0